Update WordPress cookie pattern to match defaults
The default WordPress cookies (`AUTH_COOKIE`, `SECURE_AUTH_COOKIE`,
`LOGGED_IN_COOKIE`, `TEST_COOKIE`) all start with `wordpress_`. There
are situations in which only those WP cookies are set, and none of the
others (e.g., non-admin users browing the site rather than the
dashboard). In those situations, the StackPath cache was not being
bypassed properly, and so the initial logged-in view was being cached
and delivered to other users. Additionally, logged-in users were still
seeing earlier cached versions of pages that should NOT have been coming
from the cache, now that they were logged in.

This commit assumes that the `wordpress` cookie name should have indeed
been `wordpress_*` to match the others, since there doesn't seem to be a
lone `wordpress` cookie that WP uses.

This does not address a situation where, e.g., `LOGGED_IN_COOKE` is
customized in `wp-config.php`.