Account for stackpath submenu not being set
If `$submenu['stackpath']` is not set at all, the `array_shift()` call
causes a fatal error in PHP 8+. In PHP versions prior to 8, this
situation was nonfatal.

This key could be not set due to permissions that a non-admin user might
lack, and/or permissions modified by a plugin or custom code.

So, we check to be sure that the key is set and is an array before
trying to `array_shift()` it.